application/backend/project/requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
django-allauth==0.57.2
docutils==0.21.1
googlemaps==4.10.0
gunicorn==21.2.0
numpy==1.26.4
oauthlib==3.2.2
pillow==10.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
